# Conflicts:
#	server/server-service/src/main/java/cn/torna/service/DocInfoService.java

+++ server/server-common/src/main/java/cn/torna/common/support/BaseService.java
@@ -152,6 +152,9 @@ public abstract class BaseService<E, Mapper extends CrudMapper<E, Long>> {
      * @return 返回影响行数
      */
     public int delete(E entity) {
+        if (entity == null) {
+            return 0;
+        }
         return mapper.delete(entity);
     }
